refactor: convert context to TypeScript
