--- src/ReduxUIStoreContext.js
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const ReduxUIStoreContext =  React.createContext('store');
