+++ src/index.js
@@ -2,7 +2,7 @@
 
 import ui from './ui';
 import reducer from './action-reducer';
-import {ReduxUIStoreContext} from './ReduxUIStoreContext';
+import {ReduxUIStoreContext} from './context';
 
 export default ui;
 export { reducer, ReduxUIStoreContext };
